# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Company AI Assistant API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: gemini
    description: Gemini AI chat and image operations
  - name: documents
    description: Company knowledge base document management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /gemini/conversations:
    get:
      operationId: listGeminiConversations
      tags: [gemini]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeminiConversation"
    post:
      operationId: createGeminiConversation
      tags: [gemini]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGeminiConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiConversation"
  /gemini/conversations/{id}:
    get:
      operationId: getGeminiConversation
      tags: [gemini]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiError"
    delete:
      operationId: deleteGeminiConversation
      tags: [gemini]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiError"
    patch:
      operationId: renameGeminiConversation
      tags: [gemini]
      summary: Rename a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenameGeminiConversationBody"
      responses:
        "200":
          description: Updated conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiConversation"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiError"
  /gemini/conversations/{id}/messages:
    get:
      operationId: listGeminiMessages
      tags: [gemini]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeminiMessage"
    post:
      operationId: sendGeminiMessage
      tags: [gemini]
      summary: Send a message and receive an AI response (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendGeminiMessageBody"
      responses:
        "200":
          description: SSE stream of assistant response chunks
          content:
            text/event-stream: {}
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List all uploaded documents
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload a document to the knowledge base
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                name:
                  type: string
              required:
                - file
      responses:
        "201":
          description: Uploaded document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiError"
  /documents/{id}:
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document from the knowledge base
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiError"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    GeminiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    GeminiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    CreateGeminiConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    RenameGeminiConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    SendGeminiMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    GeminiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/GeminiMessage"
      required:
        - id
        - title
        - createdAt
        - messages
    Document:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        filename:
          type: string
        size:
          type: integer
        mimeType:
          type: string
        chunkCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - filename
        - size
        - mimeType
        - chunkCount
        - createdAt
    GeminiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error