# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Company AI Assistant API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all conversations
 */
export const ListGeminiConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.date(),
});
export const ListGeminiConversationsResponse = zod.array(
  ListGeminiConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateGeminiConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get conversation with messages
 */
export const GetGeminiConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetGeminiConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.date(),
    }),
  ),
});

/**
 * @summary Delete a conversation
 */
export const DeleteGeminiConversationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Rename a conversation
 */
export const RenameGeminiConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const RenameGeminiConversationBody = zod.object({
  title: zod.string(),
});

export const RenameGeminiConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.date(),
});

/**
 * @summary List messages in a conversation
 */
export const ListGeminiMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListGeminiMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.date(),
});
export const ListGeminiMessagesResponse = zod.array(
  ListGeminiMessagesResponseItem,
);

/**
 * @summary Send a message and receive an AI response (SSE stream)
 */
export const SendGeminiMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendGeminiMessageBody = zod.object({
  content: zod.string(),
});

/**
 * @summary List all uploaded documents
 */
export const ListDocumentsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  filename: zod.string(),
  size: zod.number(),
  mimeType: zod.string(),
  chunkCount: zod.number(),
  createdAt: zod.date(),
});
export const ListDocumentsResponse = zod.array(ListDocumentsResponseItem);

/**
 * @summary Upload a document to the knowledge base
 */
export const UploadDocumentBody = zod.object({
  file: zod.instanceof(File),
  name: zod.string().optional(),
});

/**
 * @summary Delete a document from the knowledge base
 */
export const DeleteDocumentParams = zod.object({
  id: zod.coerce.number(),
});