# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Company AI Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface Document {
  id: number;
  name: string;
  filename: string;
  size: number;
  mimeType: string;
  chunkCount: number;
  createdAt: Date;
}